Baseload stuff
I don;t think this is giving the right baseload.
Most likely the problem is me setting values to zero where I dont have the info.
But, at this stage the voltage on supply point 1c is 5 times that of 1a...probably not how it should be.
Will think about this later.